* Renamed RPCManager -> RPCContainer * Moved RPCContainer implementation from SpatialSender to Utils/RPCContainer
